feat: add music context
Just like other context objects, it is global and accessible to all
consumers. In this situation, I am animating `box-shadow` of card when
music is playing. I think animating `box-shadow` is not a good idea, in
this case, I might change it later on, but for now it can be toggled
from animation toggler as well.